Flow, security and error handling improvements

+++ server.py
@@ -1,10 +1,11 @@
-from flask import Flask, render_template, request, Response, session, url_for
+from flask import Flask, render_template, request, Response, session, redirect
 from gpiozero import LED
 from signal import pause
 import datetime, os, signal
 import gevent
 import gevent.monkey
 import subprocess as sp
+import mysql.connector
 from gevent.pywsgi import WSGIServer
 
 images = os.path.join('static', 'images')
@@ -15,10 +16,36 @@ app.config['image_folder'] = images
 ledRed = LED(21)
 ledGreen = LED(20)
 
+@app.route("/", methods=['POST', 'GET'])
+def login():
+	session['user'] = None
+	u, pw,h,db = 'root', 'dmitiot', 'localhost', 'FaceReko'
+	con = mysql.connector.connect(user=u,password=pw,host=h,database=db)
+	print("Database successfully connected")
+	cur = con.cursor()
+	query = "SELECT Username, Password FROM Login"
+	cur.execute(query)
+	if request.method == 'POST':
+		for (Username, Password) in cur:
+			if request.form['user'] == Username and request.form['pass']==Password:
+				session['user'] = request.form['user']
+				return redirect('/home')
 
-@app.route("/")
+	return render_template('login.html')
+
+@app.route("/logout")
+def logout():
+	session.clear()
+	return redirect('/')
+
+@app.route("/home")
+def check():
+	if session.get('user') is None:
+		return redirect('/')
+	else:
+		return chart()
+
 def chart():
-	import mysql.connector
 	u, pw,h,db = 'root', 'dmitiot', 'localhost', 'FaceReko'
 	data = []
 	con = mysql.connector.connect(user=u,password=pw,host=h,database=db)
@@ -44,36 +71,51 @@ def chart():
 	#print('Rendered')
 
 @app.route("/activate/")
+def check2():
+	if session.get('user') is None:
+		return redirect('/')
+	else:
+		return activate()
+
 def activate():
 	extProc = sp.Popen(['python', 'rfid.py'])
 	pid = extProc.pid
 	session['proc'] = pid
 	session['status'] = 'Active'
-	templateData = {
-	      		'title' : 'FaceReko',
-	      		'response' : 'Facial Recognition active'
-   			}
-	return render_template('trigger.html', **templateData)
+	return redirect('/home')
 
 @app.route("/deactivate/")
+def check3():
+	if session.get('user') is None:
+		return redirect('/')
+	else:
+		return deactivate()
+
 def deactivate():
 	pid = session.get('proc', None)
+	
+	if pid is None:
+		return redirect('/home')
+
 	session['status'] = 'Offline'
 	os.kill(pid, signal.SIGKILL)
 	ledRed.off()
 	ledGreen.off()
-	templateData = {
-	      		'title' : 'FaceReko',
-	      		'response' : 'Facial Recognition deactivated'
-   			}
-	return render_template('trigger.html', **templateData)
+	
+	return redirect('/home')
 
 @app.route("/<img>")
+def check4(img):
+	if session.get('user') is None:
+		return redirect('/')
+	else:
+		return display_image(img)
+
 def display_image(img):
 	full_filename = os.path.join(app.config['image_folder'], img)
 	return render_template('image.html', img = full_filename)
 
 if __name__ == '__main__':
-	app.secret_key = 'A0Zr98j/3yX R~XHH!jmN]LWX/,?RT' #Hardcoded for simplicity sake
+	app.secret_key = os.urandom(30)
 	app.run(debug=True,host='0.0.0.0')
